File Displayer
The FileDisplay class contains several methods that will allow you to display the first five lines of a file, the whole contents of a file, or the whole contents of the file in numbered list form. The FileDisplayRun class demonstrates the use of loops and method calls to display the desired information.